# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arabic POS System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/dashboard:
    get:
      operationId: getDashboardStats
      tags: [reports]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily sales report
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly sales report
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/top-products:
    get:
      operationId: getTopProducts
      tags: [reports]
      summary: Get top selling products
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
        - total
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer]
        notes:
          type: string
        status:
          type: string
          enum: [completed, cancelled, refunded]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - invoiceNumber
        - items
        - subtotal
        - discount
        - tax
        - total
        - paymentMethod
        - status
        - createdAt
        - updatedAt
...
    DashboardStats:
      type: object
      properties:
        todaySales:
          type: number
        todayTransactions:
          type: integer
        monthSales:
          type: number
        monthTransactions:
          type: integer
        totalProducts:
          type: integer
        lowStockProducts:
          type: integer
        totalCustomers:
          type: integer
        totalUsers:
          type: integer
        recentSales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        salesChartData:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              total:
                type: number
              count:
                type: integer
            required:
              - date
              - total
              - count
      required:
        - todaySales
        - todayTransactions
        - monthSales
        - monthTransactions
        - totalProducts
        - lowStockProducts
        - totalCustomers
        - totalUsers
        - recentSales
        - salesChartData
    DailyReport:
      type: object
      properties:
        date:
          type: string
        totalSales:
          type: number
        totalTransactions:
          type: integer
        totalProfit:
          type: number
        averageOrderValue:
          type: number
        topProducts:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              productName:
                type: string
              quantitySold:
                type: integer
              revenue:
                type: number
            required:
              - productId
              - productName
              - quantitySold
              - revenue
        salesByHour:
          type: array
          items:
            type: object
            properties:
              hour:
                type: integer
              total:
                type: number
              count:
                type: integer
            required:
              - hour
              - total
              - count
      required:
        - date
        - totalSales
        - totalTransactions
        - totalProfit
        - averageOrderValue
        - topProducts
        - salesByHour
    MonthlyReport:
      type: object
      properties:
        year:
          type: integer
        month:
          type: integer
        totalSales:
          type: number
        totalTransactions:
          type: integer
        totalProfit:
          type: number
        averageOrderValue:
          type: number
        topProducts:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              productName:
                type: string
              quantitySold:
                type: integer
              revenue:
                type: number
            required:
              - productId
              - productName
              - quantitySold
              - revenue
        salesByDay:
          type: array
          items:
            type: object
            properties:
              day:
                type: integer
              total:
                type: number
              count:
                type: integer
            required:
              - day
              - total
              - count
      required:
        - year
        - month
        - totalSales
        - totalTransactions
        - totalProfit
        - averageOrderValue
        - topProducts
        - salesByDay
    TopProductsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              productName:
                type: string
              categoryName:
                type: string
              quantitySold:
                type: integer
              revenue:
                type: number
              profit:
                type: number
            required:
              - productId
              - productName
              - quantitySold
              - revenue
              - profit
      required:
        - data
...